chore(*): initial setup
Set up a skeleton project for a node module to be submitted to npmjs.com
There is no functionality in this release; it only contains boilerplate
code in order to hit the ground running on future projects. Test for
JavaScript code style (Standard), unit tests, and code coverage.

Add `.editorconfig` to ensure that all code submissions are uniform
Add `.gitignore` to ensure that unnecessary files do not get committed
Add `package.json` to prepare this module for npmjs.com
Add Codeship, Coveralls, and Dependency status badges to `README.md`

Closes #2