feat(gulp-pug-linter): create plugin (#6)
Create the plugin to lint Pug errors. Make sure that it works by
reading local configuration files using the existing `pug-lint` CLI.
Ignore `Stream` files. Check `Buffer` files. When a lint error is
encountered, decorate the file with an array of `.pugLinter.errors`.

Create a helper method to report on previously-encountered errors.
Once errors are encountered, log them using the basic `gulp-util`
logger. If the reporter is set to fail, throw a `PluginError` to break
the stream. Otherwise, continue the stream uninterrupted.

Ensure that every line of code is covered with unit tests.
Annotate functions where it makes sense. Write a friendly `README`
complete with installation instructions and usage examples.