feat(reporter): allow custom and external reporters
Format error message on the reporter and provide it with data.
Add support for custom reporters while maintaining 'fail' as a viable
option so that continuous integration environments break as expected.
In addition, allow users to load reporters from modules, or by passing
a reporter function into this plugin.

BREAKING CHANGE:
Though the documentation directs using 'fail' as a reporter name to use
the fail reporter, any non-null value would trigger the fail reporter;
with this modification, any value other than null or 'fail' with throw
an invalid reporter error.