feat(package): create custom commit messages for Greenekeeper
Greenekeeper used to fail the `gitlint` test because its commit messages
did not include a commit message body in them. As a result, Greenekeeper
would erroneously create new issues blaming the dependencies instead.

This commit potentially fixes that issue by adding a commit message body
after a couple of linebreaks. Potentially as opposed to actually because
it is difficult to test until the next dependencies falls out of date.